requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
urllib3>=2.0.0